chore: move Voiture Noire views to a folder
