+++ voiture_noire/views/__init__.py
@@ -1,0 +1,2 @@
+from .member_profiles_views import *
+from .other_views import *
